requirements.txt
asgiref==3.7.2
Django==4.2.4
sqlparse==0.4.4

qrcode~=7.4.2
Pillow~=10.0.0